upstage-nlp-summarization-nlp-03/requirements.txt
accelerate==0.27.2
bitsandbytes==0.42.0
datasets==2.18.0
hydra-core==1.3.2
langchain-cohere==0.1.2
langchain-community==0.0.28
langchain-core==0.1.42
omegaconf==2.3.0
pandas==2.1.4
peft==0.9.0
python-dotenv
torch==2.2.0
tqdm==4.66.1
transformers==4.38.2
trl==0.7.11
wandb==0.17.0